rust compiler: refine index handling

+++ tests/machine/x/rust/list_assign.rs
@@ -1,5 +1,5 @@
 fn main() {
     let mut nums = vec![1, 2];
-    nums[1 as usize] = 3;
-    println!("{}", nums[1 as usize]);
+    nums[1] = 3;
+    println!("{}", nums[1]);
 }
